+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - media key support for Windows and macOS
 - App is now available in Italian (thanks to Connect_Thanks2613)
 
+**Fixes**:
+- Vibration setting now available for Zwift Ride devices
+
 ### 4.2.0 (20-12-2025)
 
 BikeControl now offers a free trial period of 5 days for all features, so you can test everything before deciding to purchase a license. Please contact the support if you experience any issues!